// nightrider/src/sre-api/raindrop.gen.ts
// Code generated by raindrop build, DO NOT EDIT.
import { ActorNamespace, Ai, Annotation, App, Logger, MRNObject, SmartBucket, SmartMemory, SmartSql, Tracer,  } from '@liquidmetal-ai/raindrop-framework';

export interface Env {
  _raindrop: {
    app: App;
  };
  AI: Ai;
  annotation: Annotation<Omit<MRNObject, 'type' | 'applicationName' | 'versionId'>>;
  CONVERSATION_MEMORY: SmartMemory;
  INCIDENT_DATA: ActorNamespace<import('../incident-data/index.js').IncidentData>;
  INCIDENTS_DATABASE: SmartSql;
  KNOWLEDGE_BASE: SmartBucket;
  logger: Logger;
  tracer: Tracer;
}